Breakout DbtContext into its own module
