+++ sqlmesh/dbt/seed.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 from sqlmesh.core.model import Model, SeedKind, create_seed_model
 from sqlmesh.dbt.basemodel import BaseModelConfig
-from sqlmesh.dbt.common import DbtContext, SqlStr
+from sqlmesh.dbt.common import SqlStr
+from sqlmesh.dbt.dbtcontext import DbtContext
 
 
 class SeedConfig(BaseModelConfig):
